+++ src/pixel.rs
@@ -183,7 +183,7 @@ impl ChromaSampling {
     /// The plane dimensions must also be rounded up to accommodate odd luma plane
     ///  sizes.
     /// Cs400 returns None, as there are no chroma planes.
-    pub const fn get_decimation(self) -> Option<(usize, usize)> {
+    pub const fn get_decimation(self) -> Option<(u32, u32)> {
         use self::ChromaSampling::*;
         match self {
             Cs420 => Some((1, 1)),
@@ -194,11 +194,7 @@ impl ChromaSampling {
     }
 
     /// Calculates the size of a chroma plane for this sampling type, given the luma plane dimensions.
-    pub const fn get_chroma_dimensions(
-        self,
-        luma_width: usize,
-        luma_height: usize,
-    ) -> (usize, usize) {
+    pub const fn get_chroma_dimensions(self, luma_width: u32, luma_height: u32) -> (u32, u32) {
         if let Some((ss_x, ss_y)) = self.get_decimation() {
             ((luma_width + ss_x) >> ss_x, (luma_height + ss_y) >> ss_y)
         } else {
